Only make external call to freegeoip.net if location data required
Amber has some geo-specific features that allow the administrator to
configure different behavior based on the visitor's country. These
features require a call from the user's browser to an external service
(freegeoip.net) which tells us what country the user is in. This
service should only be invoked if it would be required on the page
being viewed. (#28)