Add error message if required PHP cURL extension isn't installed
The PHP cURL extension is required for Amber to retrieve and cache
snapshots of linked sites. Currently the Amber WordPress plugin
silently fails at making snapshots, logging “CURL not installed” errors
that may be difficult for end users to find and debug.

This checks if the curl_init function exists, indicating that the cURL
extension is installed and working. If it is not, an error message is
displayed with a link with installation instructions from PHP.net.

git-svn-id: http://plugins.svn.wordpress.org/amberlink/trunk@1501677 b8457f37-d9ea-0310-8a92-e5e31aec5664