Prevent access to snapshots except through Amber iframe
If a snapshot is loaded directly, without a surrounding iframe that
puts it in a sandbox, the viewer could be at risk of an XSS attack.
Prevent this by (a) checking that views of cached content have a
referrer header that matches that of the expected containing iframe;
and (b) restricting access to the wp-content/uploads/amber directory
by placing a restrictive .htaccess fiels in that directory. (#34)

git-svn-id: http://plugins.svn.wordpress.org/amberlink/trunk@1501703 b8457f37-d9ea-0310-8a92-e5e31aec5664